requirements.txt
langchain
langchain-openai
langchain-ollama
chromadb
pandas
pydantic
python-dotenv
beautifulsoup4
streamlit
streamlit-chat